Refactor admin.py into multiple files
